Split 'ruby-pg' module and 'postgres' module into separate parts of the repository. They will be maintained separately, and be released as separate gems from this point forward.
One reason for this change is because we need separate
gems. If we distribute both modules as one gem, the
documentation is not properly generated due to class name
conflicts.

Another reason is to reduce confusion between the two
modules.

Also, I don't plan on making many improvements (aside
from bugfixes and portability issues) to the old
'postgres' code. I'd like to get it to a stable state,
and leave it alone. Meanwhile, the 'pg' module can be
developed on a faster timeline.