Fix PGconn#async_exec and PGconn#get_last_result to properly clear all results when an exception is raised. Before, the connection could be left in a state that is not ready for new commands.
Also, change PGconn#get_result	to not raise an	exception when the
result is in an error state. You can still check the result object for
an error, but it's not appropriate to raise an exception, because that
could prevent the caller from properly clearing all results.

Thanks to Tarmo Tänav.